move eval datasets and remove data_path in config
