+++ nemo_rl/data/datasets/eval_datasets/local_math_dataset.py
@@ -14,6 +14,7 @@
 
 """Local math dataset."""
 
+import os
 from typing import Any, Literal, Optional
 
 from datasets import load_dataset
@@ -25,23 +26,22 @@ from nemo_rl.data.interfaces import TaskDataSpec
 class LocalMathDataset:
     def __init__(
         self,
-        data_paths: str | list[str],
+        data_path: str,
         problem_key: str,
         solution_key: str,
-        name: str,
         split: Optional[str] = None,
         file_format: Literal["csv", "json"] = "csv",
         prompt_file: Optional[str] = None,
         system_prompt_file: Optional[str] = None,
     ):
-        ds = load_dataset(file_format, data_files=data_paths)
+        ds = load_dataset(file_format, data_files=data_path)
         if split is not None:
             ds = ds[split]
         self._problem_key = problem_key
         self._solution_key = solution_key
         self.rekeyed_ds = ds.map(self._rekey, remove_columns=ds.column_names)
         self.task_spec = TaskDataSpec(
-            task_name=name,
+            task_name=os.path.basename(data_path).split(".")[0],
             prompt_file=prompt_file,
             system_prompt_file=system_prompt_file,
         )
